requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
coverage==5.4
Django==3.1.5
factory-boy==3.2.0
Faker==6.5.0
glob2==0.7
inflection==0.5.1
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
parse==1.19.0
parse-type==0.5.2
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-bdd==4.0.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
pytz==2021.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
text-unicode==1.3
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
django-filter~=2.4.0